pytest with postgres creation is working
