--- .github/workflows/pytest.yml
@@ -1,48 +1,0 @@
-name: Pytest with PostgreSQL Service
-on: [push]
-
-jobs:
-  runner-job:
-    runs-on: ubuntu-latest
-
-    # Service containers to run with `runner-job`
-    services:
-      # Label used to access the service container
-      postgres:
-        # Docker Hub image
-        image: postgres:12
-        # Provide the password for postgres
-        env:
-          POSTGRES_PASSWORD: postgres
-          POSTGRES_DB: test_service_providers
-        # Set health checks to wait until postgres has started
-        options: >-
-          --health-cmd pg_isready
-          --health-interval 10s
-          --health-timeout 5s
-          --health-retries 5
-        ports:
-          # Maps tcp port 5432 on service container to the host
-          - 5432:5432
-
-    steps:
-      # Downloads a copy of the code in your repository before running CI tests
-      - name: Check out repository code
-        uses: actions/checkout@v2
-      - name: check python version
-        run: python --version
-      - name: Install dependencies
-        run: |
-          python3 -m pip install --upgrade pip
-          pip3 install pytest setuptools
-          if [ -f requirements.txt ]; then pip3 install -r requirements.txt; fi
-      - name: run tests
-        
-        run: python3 -m pytest -v
-        # Environment variable used by the `client.js` script to create
-        # a new PostgreSQL client.
-        env:
-          # The hostname used to communicate with the PostgreSQL service container
-          POSTGRES_HOST: localhost
-          SQLALCHEMY_TRACK_MODIFICATIONS: False
-          POSTGRES_PORT: 5432
