+++ .github/workflows/pytest2.yml
@@ -1,4 +1,4 @@
-name: PYTEST2
+name: PYTEST
 on: [push]
 
 jobs:
@@ -29,7 +29,7 @@ jobs:
       run: |
         pip install --upgrade pip
         pip install -r requirements.txt       
-        python -m pytest tests_new
+        python -m pytest tests_new -vvv
       env:
         TEST_DATABASE_URL: postgresql://postgres:postgres@localhost:5432/test_service_providers
         POSTGRES_VER: 11
